requirements.txt
termcolor==1.1.0
progressbar2==3.18.0